update permission mutation

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 const { makeANiceEmail, transport } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
@@ -49,9 +50,15 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id title}`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
     // 2. Check if they own that item, or have the permissions
-    // TODO
+    const isOwner = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => {
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    });
+    if (!isOwner && !hasPermissions) {
+      throw new Error('You are not allowed to delete');
+    }
     // 3. Delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -172,6 +179,38 @@ const Mutations = {
     });
     // Finalllllly we return the user to the browser
     return user;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // check if login
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in!');
+    }
+    //query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId,
+        },
+      },
+      info
+    );
+
+    // check if has user permission
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+    // update the user permission
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions,
+          },
+        },
+        where: {
+          id: args.userId,
+        },
+      },
+      info
+    );
   }
 };
